# packages/api-types/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "JellyfinSuite Plugin API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8096",
      "description": "Jellyfin server (default port)"
    }
  ],
  "security": [],
  "paths": {
...
    "/JellyfinSuite/SeekPreview/{itemId}": {
      "get": {
        "tags": [
          "SeekPreview"
        ],
        "operationId": "seekPreview_getFrame",
        "summary": "Get Frame",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "positionMs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "prefetch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG frame",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/JellyfinSuite/SeekPreview/{itemId}/frame-info": {
      "get": {
        "tags": [
          "SeekPreview"
        ],
        "operationId": "seekPreview_getFrameInfo",
        "summary": "Get Frame Info",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "positionMs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Frame info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameInfoDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/JellyfinSuite/SeekPreview/{itemId}/frame-index": {
      "get": {
        "tags": [
          "SeekPreview"
        ],
        "operationId": "seekPreview_getFrameIndex",
        "summary": "Get Frame Index",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Frame index",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameIndexDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FrameInfoDto": {
        "type": "object",
        "properties": {
          "frameIdx": {
            "type": "integer",
            "format": "int64"
          },
          "frameStartMs": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "frameIdx",
          "frameStartMs"
        ]
      },
      "FrameIndexDto": {
        "type": "object",
        "properties": {
          "frames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FrameIndexEntryDto"
            }
          },
          "fps": {
            "$ref": "#/components/schemas/FpsFracDto"
          }
        },
        "required": [
          "frames",
          "fps"
        ]
      },
      "FrameIndexEntryDto": {
        "type": "object",
        "properties": {
          "frameIndex": {
            "type": "integer",
            "format": "int64"
          },
          "ms": {
            "type": "integer",
            "format": "int64"
          },
          "isKey": {
            "type": "boolean"
          }
        },
        "required": [
          "frameIndex",
          "ms",
          "isKey"
        ]
      },
      "FpsFracDto": {
        "type": "object",
        "properties": {
          "num": {
            "type": "integer",
            "format": "int64"
          },
          "den": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "num",
          "den"
        ]
      },
...
    }
  }
}