# packages/api-types/src/model/frameInfoDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * JellyfinSuite Plugin API
 * OpenAPI spec version: 1.0.0
 */

export interface FrameInfoDto {
  frameIdx: number;
  frameStartMs: number;
}